1.修改名称、版本

+++ lib/src/models/candle_sticks_style.dart
@@ -25,7 +25,7 @@ class CandleSticksStyle {
 
   final Color toolBarColor;
 
-  final int borderTextSize;
+  final double borderTextSize;
 
   CandleSticksStyle({
     required this.borderColor,
